+++ backend/factory-management/pom.xml
@@ -8,7 +8,7 @@
         <version>3.0.2</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
-    <groupId>com.projetada</groupId>
+    <groupId>com.projedata</groupId>
     <artifactId>factory-management</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>factory-management</name>
