+++ backend/factory-management/src/main/java/com/projedata/factorymanagement/FactoryManagementApplication.java
@@ -1,0 +1,13 @@
+package com.projedata.factorymanagement;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class FactoryManagementApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(FactoryManagementApplication.class, args);
+	}
+
+}
