+++ backend/factory-management/src/main/java/com/projedata/factorymanagement/config/ModelMapperConfig.java
@@ -1,0 +1,15 @@
+package com.projedata.factorymanagement.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ModelMapperConfig {
+
+    @Bean
+    public ModelMapper mapper() {
+        return new ModelMapper();
+    }
+
+}
