+++ backend/factory-management/src/main/java/com/projedata/factorymanagement/exceptions/DataIntegrityViolationException.java
@@ -1,0 +1,8 @@
+package com.projedata.factorymanagement.exceptions;
+
+public class DataIntegrityViolationException extends RuntimeException {
+
+    public DataIntegrityViolationException(String message) {
+        super(message);
+    }
+}
