+++ backend/factory-management/src/main/java/com/projedata/factorymanagement/exceptions/ObjectNotFoundException.java
@@ -1,0 +1,8 @@
+package com.projedata.factorymanagement.exceptions;
+
+public class ObjectNotFoundException extends RuntimeException {
+
+    public ObjectNotFoundException(String message) {
+        super(message);
+    }
+}
