+++ backend/factory-management/src/main/java/com/projedata/factorymanagement/exceptions/StandardError.java
@@ -1,0 +1,50 @@
+package com.projedata.factorymanagement.exceptions;
+
+import java.time.LocalDateTime;
+
+public class StandardError {
+
+    private LocalDateTime timestamp;
+    private Integer status;
+    private String error;
+    private String path;
+
+    public StandardError(LocalDateTime timestamp, Integer status, String error, String path) {
+        this.timestamp = timestamp;
+        this.status = status;
+        this.error = error;
+        this.path = path;
+    }
+
+    public LocalDateTime getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(LocalDateTime timestamp) {
+        this.timestamp = timestamp;
+    }
+
+    public Integer getStatus() {
+        return status;
+    }
+
+    public void setStatus(Integer status) {
+        this.status = status;
+    }
+
+    public String getError() {
+        return error;
+    }
+
+    public void setError(String error) {
+        this.error = error;
+    }
+
+    public String getPath() {
+        return path;
+    }
+
+    public void setPath(String path) {
+        this.path = path;
+    }
+}
