+++ backend/factory-management/src/main/java/com/projedata/factorymanagement/models/Material.java
@@ -1,0 +1,71 @@
+package com.projedata.factorymanagement.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import jakarta.persistence.*;
+
+import java.io.Serializable;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.UUID;
+
+@Entity
+@Table(name = "TB_MATERIAL")
+public class Material implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private UUID id;
+    @Column(nullable = false, length = 100)
+    private String name;
+    @Column(nullable = false)
+    private Integer stock;
+
+    @JsonIgnore
+    @ManyToMany(mappedBy = "materials")
+    private Set<Product> products = new HashSet<>();
+
+    public Material() {
+    }
+
+    public Material(UUID id, String name, Integer stock, Set<Product> products) {
+        this.id = id;
+        this.name = name;
+        this.stock = stock;
+        this.products = products;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public void setId(UUID id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Integer getStock() {
+        return stock;
+    }
+
+    public void setStock(Integer stock) {
+        this.stock = stock;
+    }
+
+    public Set<Product> getProducts() {
+        return products;
+    }
+
+    public void setProducts(Set<Product> products) {
+        this.products = products;
+    }
+}
+
