+++ backend/factory-management/src/main/java/com/projedata/factorymanagement/models/Product.java
@@ -1,0 +1,75 @@
+package com.projedata.factorymanagement.models;
+
+import jakarta.persistence.*;
+
+import java.io.Serializable;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.UUID;
+
+
+@Entity
+@Table(name = "TB_PRODUCT")
+public class Product implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private UUID id;
+    @Column(nullable = false, length = 100)
+    private String name;
+    @Column(nullable = false)
+    private Double value;
+
+    @ManyToMany
+    @JoinTable(
+            name = "TB_RECIPE",
+            joinColumns = {@JoinColumn(name = "id_product", referencedColumnName = "id")},
+            inverseJoinColumns = {@JoinColumn(name = "id_material", referencedColumnName = "id")}
+    )
+    private Set<Material> materials = new HashSet<>();
+
+    public Product() {
+    }
+
+    public Product(UUID id, String name, Double value, Set<Material> materials) {
+        this.id = id;
+        this.name = name;
+        this.value = value;
+        this.materials = materials;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public void setId(UUID id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public Double getValue() {
+        return value;
+    }
+
+    public void setValue(Double value) {
+        this.value = value;
+    }
+
+    public Set<Material> getMaterials() {
+        return materials;
+    }
+
+    public void setMaterials(Set<Material> materials) {
+        this.materials = materials;
+    }
+
+}
