+++ backend/factory-management/src/main/java/com/projedata/factorymanagement/repositories/ProductRepository.java
@@ -1,0 +1,13 @@
+package com.projedata.factorymanagement.repositories;
+
+import com.projedata.factorymanagement.models.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, UUID> {
+    Optional<Product> findByName(String name);
+}
