+++ backend/factory-management/src/main/java/com/projedata/factorymanagement/services/MaterialService.java
@@ -1,0 +1,20 @@
+package com.projedata.factorymanagement.services;
+
+import com.projedata.factorymanagement.dto.MaterialDto;
+import com.projedata.factorymanagement.models.Material;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface MaterialService {
+
+    Material findById(UUID id);
+
+    List<Material> findAll();
+
+    Material create(MaterialDto materialDto);
+
+    Material update(MaterialDto materialDto);
+
+    void delete(UUID id);
+}
