--- backend/factory-management/src/main/java/com/projetada/factorymanagement/FactoryManagementApplication.java
@@ -1,13 +1,0 @@
-package com.projetada.factorymanagement;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class FactoryManagementApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(FactoryManagementApplication.class, args);
-	}
-
-}
