--- backend/factory-management/src/main/java/com/projetada/factorymanagement/config/ModelMapperConfig.java
@@ -1,15 +1,0 @@
-package com.projetada.factorymanagement.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ModelMapperConfig {
-
-    @Bean
-    public ModelMapper mapper() {
-        return new ModelMapper();
-    }
-
-}
