--- backend/factory-management/src/main/java/com/projetada/factorymanagement/dto/MaterialDto.java
@@ -1,61 +1,0 @@
-package com.projetada.factorymanagement.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-
-import com.projetada.factorymanagement.models.Product;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.Set;
-import java.util.UUID;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class MaterialDto {
-
-    private UUID id;
-
-    private String name;
-
-    private Integer stock;
-
-    @JsonIgnore
-    private Set<Product> products;
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Integer getStock() {
-        return stock;
-    }
-
-    public void setStock(Integer stock) {
-        this.stock = stock;
-    }
-
-    public Set<Product> getProducts() {
-        return products;
-    }
-
-    public void setProducts(Set<Product> products) {
-        this.products = products;
-    }
-
-}
