--- backend/factory-management/src/main/java/com/projetada/factorymanagement/dto/ProductDto.java
@@ -1,59 +1,0 @@
-package com.projetada.factorymanagement.dto;
-
-import com.projetada.factorymanagement.models.Material;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.Set;
-import java.util.UUID;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ProductDto {
-
-    private UUID id;
-
-    private String name;
-
-    private Double value;
-
-    private Set<Material> materials;
-
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Double getValue() {
-        return value;
-    }
-
-    public void setValue(Double value) {
-        this.value = value;
-    }
-
-    public Set<Material> getMaterials() {
-        return materials;
-    }
-
-    public void setMaterials(Set<Material> materials) {
-        this.materials = materials;
-    }
-
-}
