--- backend/factory-management/src/main/java/com/projetada/factorymanagement/exceptions/DataIntegrityViolationException.java
@@ -1,8 +1,0 @@
-package com.projetada.factorymanagement.exceptions;
-
-public class DataIntegrityViolationException extends RuntimeException {
-
-    public DataIntegrityViolationException(String message) {
-        super(message);
-    }
-}
