--- backend/factory-management/src/main/java/com/projetada/factorymanagement/exceptions/ObjectNotFoundException.java
@@ -1,8 +1,0 @@
-package com.projetada.factorymanagement.exceptions;
-
-public class ObjectNotFoundException extends RuntimeException {
-
-    public ObjectNotFoundException(String message) {
-        super(message);
-    }
-}
