--- backend/factory-management/src/main/java/com/projetada/factorymanagement/exceptions/ResourceExceptionHandler.java
@@ -1,39 +1,0 @@
-package com.projetada.factorymanagement.exceptions;
-
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.time.LocalDateTime;
-
-@ControllerAdvice
-public class ResourceExceptionHandler {
-
-    @ExceptionHandler(ObjectNotFoundException.class)
-    public ResponseEntity<StandardError> objectNotFoundException(
-            ObjectNotFoundException exception, HttpServletRequest request) {
-
-        StandardError error = new StandardError(
-                LocalDateTime.now(),
-                HttpStatus.NOT_FOUND.value(),
-                exception.getMessage(),
-                request.getRequestURI());
-
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(error);
-    }
-    @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<StandardError> dataIntegrityViolationException(
-            DataIntegrityViolationException exception, HttpServletRequest request) {
-
-        StandardError error = new StandardError(
-                LocalDateTime.now(),
-                HttpStatus.BAD_REQUEST.value(),
-                exception.getMessage(),
-                request.getRequestURI());
-
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(error);
-    }
-
-}
