--- backend/factory-management/src/main/java/com/projetada/factorymanagement/repositories/MaterialRepository.java
@@ -1,14 +1,0 @@
-package com.projetada.factorymanagement.repositories;
-
-import com.projetada.factorymanagement.models.Material;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface MaterialRepository extends JpaRepository<Material, UUID> {
-
-    Optional<Material> findByName(String name);
-}
