--- backend/factory-management/src/main/java/com/projetada/factorymanagement/repositories/ProductRepository.java
@@ -1,14 +1,0 @@
-package com.projetada.factorymanagement.repositories;
-
-import com.projetada.factorymanagement.models.Material;
-import com.projetada.factorymanagement.models.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, UUID> {
-    Optional<Product> findByName(String name);
-}
