+++ backend/factory-management/src/test/java/com/projedata/factorymanagement/FactoryManagementApplicationTests.java
@@ -1,0 +1,10 @@
+package com.projedata.factorymanagement;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class FactoryManagementApplicationTests {
+
+
+}
