+++ backend/factory-management/src/test/java/com/projedata/factorymanagement/config/ContainersEnvironment.java
@@ -1,0 +1,13 @@
+package com.projedata.factorymanagement.config;
+
+import com.projedata.factorymanagement.containers.PostgresTestContainer;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@Testcontainers
+public class ContainersEnvironment {
+
+    @Container
+    public static PostgreSQLContainer postgreSQLContainer = PostgresTestContainer.getInstance();
+}
