+++ backend/factory-management/src/test/java/com/projedata/factorymanagement/containers/PostgresTestContainer.java
@@ -1,0 +1,33 @@
+package com.projedata.factorymanagement.containers;
+
+import org.testcontainers.containers.PostgreSQLContainer;
+
+public class PostgresTestContainer extends PostgreSQLContainer<PostgresTestContainer> {
+
+    public static final String IMAGE_VERSION = "postgres:15.0";
+    public static final String DATABASE_NAME = "test";
+    public static PostgreSQLContainer container;
+
+    public PostgresTestContainer() {
+        super(IMAGE_VERSION);
+    }
+
+    public static PostgreSQLContainer getInstance() {
+        if (container == null) {
+            container = new PostgresTestContainer().withDatabaseName(DATABASE_NAME);
+        }
+        return container;
+    }
+
+    @Override
+    public void start() {
+        super.start();
+        System.setProperty("DB_URL", container.getJdbcUrl());
+        System.setProperty("DB_USERNAME", container.getUsername());
+        System.setProperty("DB_PASSWORD", container.getPassword());
+    }
+
+    @Override
+    public void stop() {
+    }
+}
