--- backend/factory-management/src/test/java/com/projetada/factorymanagement/FactoryManagementApplicationTests.java
@@ -1,10 +1,0 @@
-package com.projetada.factorymanagement;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class FactoryManagementApplicationTests {
-
-
-}
