--- backend/factory-management/src/test/java/com/projetada/factorymanagement/config/ContainersEnvironment.java
@@ -1,13 +1,0 @@
-package com.projetada.factorymanagement.config;
-
-import com.projetada.factorymanagement.containers.PostgresTestContainer;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-@Testcontainers
-public class ContainersEnvironment {
-
-    @Container
-    public static PostgreSQLContainer postgreSQLContainer = PostgresTestContainer.getInstance();
-}
